studyguru-auth/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
email-validator>=2.1.0
SQLAlchemy==2.0.45
psycopg[binary]
httpx==0.28.1
asyncpg==0.27.0